Ice to water melting
this simulation was performed without Boussinesq approximation thus the liquid fraction on the top portion is not larger than of the bottom portion. Left wall was is 80 degree C and the right one is 0 degree C. The interface in the png is at 80 sec of flow time. Red portion is water,blue is ice. green is somewhere between